requirements.txt
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
tqdm>=4.66
pyomo>=6.7
gurobipy>=10.0
torch>=2.1
pennylane>=0.35
